requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.40.0
langchain==1.2.10
langchain-groq==1.1.2
langchain-core==1.2.11
python-dotenv==1.2.1
pydantic==2.12.5
